# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API generating customizable fizz-buzz lists and reporting the most frequent request.",
        "title": "FizzBuzz API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/fizzbuzz": {
            "post": {
                "description": "Returns the numbers from 1 to `limit` where multiples of `int1` are replaced by `str1`, multiples of `int2` by `str2`, and multiples of both by `str1str2`.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "fizzbuzz"
                ],
                "summary": "Generate a fizz-buzz list",
                "parameters": [
                    {
                        "description": "Fizz-buzz parameters",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.FizzBuzzRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/history": {
            "get": {
                "description": "Returns successful generation requests grouped by URL path and validated parameters, newest request first. requested_at is the UTC time of the latest matching request.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stats"
                ],
                "summary": "Generation request history",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.HistoryEntry"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/history/export": {
            "get": {
                "description": "Downloads all tracked successful generation request groups as an XLSX spreadsheet.",
                "produces": [
                    "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet"
                ],
                "tags": [
                    "stats"
                ],
                "summary": "Export generation history to Excel",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/stats": {
            "get": {
                "description": "Returns the parameters of the most frequent valid request and its number of hits. On a tie, the request that reached the count first is returned.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stats"
                ],
                "summary": "Most frequent request",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.StatsResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controller.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controller.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "int1 must be a positive integer"
                }
            }
        },
        "controller.FizzBuzzRequest": {
            "type": "object",
            "required": [
                "int1",
                "int2",
                "limit",
                "str1",
                "str2"
            ],
            "properties": {
                "int1": {
                    "type": "integer",
                    "minimum": 1,
                    "example": 3
                },
                "int2": {
                    "type": "integer",
                    "minimum": 1,
                    "example": 5
                },
                "limit": {
                    "type": "integer",
                    "minimum": 1,
                    "example": 100
                },
                "str1": {
                    "type": "string",
                    "maxLength": 64,
                    "example": "fizz"
                },
                "str2": {
                    "type": "string",
                    "maxLength": 64,
                    "example": "buzz"
                }
            }
        },
        "controller.StatsResponse": {
            "type": "object",
            "properties": {
                "hits": {
                    "type": "integer",
                    "example": 42
                },
                "int1": {
                    "type": "integer",
                    "example": 3
                },
                "int2": {
                    "type": "integer",
                    "example": 5
                },
                "limit": {
                    "type": "integer",
                    "example": 100
                },
                "str1": {
                    "type": "string",
                    "example": "fizz"
                },
                "str2": {
                    "type": "string",
                    "example": "buzz"
                }
            }
        },
        "domain.HistoryEntry": {
            "type": "object",
            "properties": {
                "body": {
                    "$ref": "#/definitions/domain.Params"
                },
                "hits": {
                    "type": "integer"
                },
                "path_url": {
                    "type": "string"
                },
                "requested_at": {
                    "type": "string"
                }
            }
        },
        "domain.Params": {
            "type": "object",
            "properties": {
                "int1": {
                    "type": "integer"
                },
                "int2": {
                    "type": "integer"
                },
                "limit": {
                    "type": "integer"
                },
                "str1": {
                    "type": "string"
                },
                "str2": {
                    "type": "string"
                }
            }
        }
    }
}